extracted the `Platform` interface from the `DropwizardPlatform`.
