+++ graceland-example/src/main/java/io/graceland/example/ExampleApplication.java
@@ -1,15 +1,15 @@
 package io.graceland.example;
 
 import io.graceland.example.startingon.StartingOnCountingPlugin;
+import io.graceland.platform.DropwizardPlatform;
 import io.graceland.platform.Platform;
 import io.graceland.platform.application.SimpleApplication;
 
 public class ExampleApplication extends SimpleApplication {
 
     public static void main(String[] args) throws Exception {
-        Platform
-                .forApplication(new ExampleApplication())
-                .start(args);
+        Platform platform = DropwizardPlatform.forApplication(new ExampleApplication());
+        platform.start(args);
     }
 
     @Override
